Clean up Qt socket notifier to avoid spurious interrupt handler calls
Closes #29688

Objects without a parent are not necessarily cleaned up in `PyQt5/6`
when their reference count reaches zero, and must be explicitly cleaned
up with `deleteLater()`

This prevents the notifier firing after the signal handling was supposed
to have been reset to its previous state.

Rather than have both `bakend_bases._allow_interrupt()` and
`backend_qt._allow_interrupt_qt()` hold a reference to the notifier, we
pass it to  the backend-specific `handle_signint()` function for
cleanup.

Note the approach to cleaning up the notifier with `.deleteLater()`
followed by `sendPostedEvents()` is the documented workaround for when
immediate deletion is desired:

https://doc.qt.io/qt-6/qobject.html#deleteLater

This ensures the object is still deleted up even if the same event loop
does not run again.