Run subprocess-starting tests on separate runners
I suspect that we are running out of process IDs, or the scheduler
is not letting all subprocesses execute, leading to timeouts in
tests that start subprocesses when other long-running tests are
executing. This happens often with Python 3.14 tests for some reason.

Add a new pytest marker for the tests that start a subprocess at
top level, and run these tests in separate runners without
pytest-xdist parallelization.