Run subprocess-starting tests in a separate xdist loadgroup
I suspect that we are running out of process IDs, or the scheduler
is not letting all subprocesses execute, leading to timeouts in
tests that start subprocesses when other long-running tests are
executing. This happens often with Python 3.14 tests for some reason.

Put the tests that start a subprocess in a separate loadgroup
so that no two of them get executed in parallel.

Alternative to #30264.