Update spines.py
Pass *None* to the set_bounds() function should leave the limit unchanged, but it fails because the get_bounds() function returns None by default, which sets the *old_high* = None. 
Inspired by the codes in the _adjust_location function, the get_bounds function is now rewritten for the default self._bounds = None.