MNT: Prefer capitalized logging levels (#30339)
* MNT: Prefer capitalized logging levels

#30302 just introduced literals for the log levels. Previously, the
situation was a bit vague: While the docstring described the levels
as all lowercase, in fact any casing is accepted. This PR changes the
preferred casing to be all-capitalized in analogy to the logging
standard library, which only supports this casing. Lowercase remains
supported because it's not worth an API breakage.

We might later consider to also accept logging levels directly, i.e.
`logging.DEBUG` to be more congruent with the logging library. But this
can be done any time. It's important to get this PR in soon to not
release the typing Literal in lowercase form.

* Update lib/matplotlib/__init__.py

Co-authored-by: Thomas A Caswell <tcaswell@gmail.com>

---------

Co-authored-by: Thomas A Caswell <tcaswell@gmail.com>