ENH: Scroll to zoom
Implements a minimal version of #20317, in particular https://github
.com/matplotlib/matplotlib/pull/20317#issuecomment-2233156558:

When any of the axes manipulation tools is active (pan or zoom tool), a
mouse scroll results in a zoom towards the cursor, keeping aspect ratio.

I've decided to require an active manipulation tool, so that without any
active tool the plot cannot be changed (accidentally) - as before.
For convenience, scroll-to-zoom is allowed with both the zoom and pan
tools. Limiting further feels unnecessarily restrictive.

Zooming is also limited to not having a modifier key pressed. This is
because we might later want to add scroll+modifiers for other operations
. It's better for now not to react to these at all to not introduce
behaviors we later want to change.