Use pathlib in texmanager.
Only a few changes are needed.  Note that make_png returns an absolute
path so path joining is not even needed here.

Also restore the recently incorrectly removed usage of dpi in generating
the png filename.

Also drop an unrelated but outdated comment re: background handling,
which is obsolete since 8940c66 (look for `if hack: ...`).