Deprecate redundant axes parameter to RadialLocator.
We can just directly fetch the relevant information from
`locator.axis.axes`; the case where `locator.axis.axes` was different
from `axes` was not tested and doesn't make much sense anyways.

The change also goes in the direction of having Locators carry less
state (and hopefully, in the future, not being associated with an axis,
but rather having the axis passed as argument to `__call__`).