MNT: Pending-deprecate setting colormap extremes
Per #29141 we have the long-term plan to make colormaps immutable. As a
result, the in-place modifications have to be removed. We take this
particularly slow with at least two minor releases of pending
deprecation and two further releases of actual deprecations. As it's
quite a common concept and migrating away will take time.