MNT: Introduce cmap(..., by_index=...) parameter
This allows to explicitly state whether inputs should be handled as
normalized values [0, 1] or indices into the colormap.

The value `by_index='auto'` is the current default and makes the
interpretation dependent on the data type. Resulting in surprising
behavior #28198. It is planned to deprecate this in a follow-up PR and
switch to
`by_index=False` eventually. The standard case of float inputs will not
be affected by this change. But users that want to pass indices, will
then have to explicitly use `by_index=True`.