DOC: Discouraged duplicate colormaps
Closes #30796.

State that the use of the colormaps gist_gray, grist_yarg, binary is
discouraged. To be clear: We don't have a plan to depreacte them. Even
though I've decided to remove them from the plotting and only mention
them by name in the admonition to de-emphasize them in the docs and push
 towards the recommended alternatives.